refactor: move FoodPlanMeal, FoodPlanMessage  and MessageModal to organism components
