+++ src/components/organisms/FoodPlanMeal/foodplan.scss
@@ -1,0 +1,177 @@
+@import "../../../styles/vars";
+
+#mealGrid {
+  margin: 0;
+  padding: 0;
+  & > #mealplanSubtitle {
+    @include tagSubtitle();
+    color: $blue5;
+  }
+  & > #dailyMealRow {
+    flex-direction: column;
+    margin-bottom: 1em;
+    padding: 0 0 1em;
+    & > .oneMealColumn {
+      align-items: center;
+      border-bottom: 1px solid $blue5;
+      display: flex;
+      flex-direction: column;
+      margin-bottom: 1em;
+      padding: 0 0 2em;
+      width: 100%;
+      & > .oneMealSegment {
+        @include mealplanSegment();
+        & > .oneMealHeader {
+          color: $blue5;
+          @include mealplanHeader();
+        }
+        & > .cheatmealCheckbox {
+          @include mealplanToggle;
+          & > input {
+            &:checked + label {
+              color: $pink5 !important;
+              &::before {
+                background-color: $pink5 !important;
+              }
+            }
+          }
+        }
+      }
+      & > .oneMealForm {
+        @include formMealplan();
+        & > .mealFields {
+          & > .foodDropdown.visible > .text:not(.default) {
+            color: $green2;
+          }
+          & > .foodDropdown {
+            & > .icon {
+              color: $blue5;
+            }
+            & > .menu.transition {
+              & > .selected.item {
+                color: black;
+              }
+            }
+          }
+        }
+      }
+      & > .messagecheat {
+        animation-duration: 0.3s;
+        animation-name: opacity0to1;
+        animation-timing-function: ease-in-out;
+        font-size: 0.8em;
+        margin: auto;
+        max-width: 25em;
+        text-align: center;
+      }
+    }
+  }
+  & > #snackRow {
+    flex-direction: column;
+    padding-top: 0;
+    & > #snackColumn {
+      display: flex;
+      flex-direction: column;
+      padding: 0;
+      width: 100%;
+      & > #snackSegment {
+        @include mealplanSegment();
+        & > #snackHeader {
+          color: $yellow3;
+          @include mealplanHeader();
+        }
+        & > .cheatmealCheckbox {
+          @include mealplanToggle;
+          & > input {
+            &:checked + label {
+              color: $yellow3 !important;
+              &::before {
+                background-color: $yellow3 !important;
+              }
+            }
+          }
+        }
+      }
+      & > #snackForm {
+        @include formMealplan();
+        & > .mealFields {
+          & > .foodDropdown.visible > .text:not(.default) {
+            color: $yellow1;
+          }
+          & > .foodDropdown {
+            & > .icon {
+              color: $yellow1;
+            }
+            & > .menu.transition {
+              & > .selected.item {
+                color: $yellow1;
+              }
+            }
+          }
+        }
+      }
+      & > .snackmessage {
+        animation-duration: 0.3s;
+        animation-name: opacity0to1;
+        animation-timing-function: ease-in-out;
+        font-size: 0.8em;
+        margin: auto;
+        max-width: 25em;
+        text-align: center;
+      }
+    }
+  }
+}
+
+@media (min-width: 768px) {
+  #mealGrid {
+    padding: 1em 0;
+  }
+}
+
+@media (min-width: 1024px) {
+  #mealGrid {
+    & > #dailyMealRow {
+      border-bottom: 1px solid $blue5;
+      flex-direction: row;
+      justify-content: space-between;
+      margin: 0 auto 1em;
+      padding: 0;
+      & > .oneMealColumn {
+        border-bottom: none;
+        flex-grow: 0;
+        margin-bottom: 0;
+        width: 31%;
+        & > .oneMealSegment {
+          border-bottom: 2px solid $blue5;
+          flex-direction: column;
+          margin: 0 auto 2em;
+          & > .oneMealHeader {
+            margin: 0 0 0.8em;
+          }
+        }
+        & > .oneMealForm {
+          border-right: 1px solid $blue5;
+          margin: 0 0 0 auto;
+          & > .mealFields {
+            margin: 0 0 1em auto;
+            padding-right: 0.5em;
+          }
+          & > .mealVegAndFruit {
+            padding-right: 0.5em;
+          }
+        }
+        & > .messagecheat {
+          border-right: 1px solid $blue5;
+        }
+      }
+    }
+    & > #snackRow {
+      & > #snackColumn {
+        & > #snackSegment {
+          border-bottom: 2px solid $yellow3;
+        }
+      }
+    }
+  }
+}
